Added CSS rules for the various levels of opacity
Added a CSS rule for each level of opacity from 0.0 to 1.0 going from a to k. Letters had to be used as numbers in CSS class rules cant be used or I don't know a way of making it so it can.

Because of CSS inability to have numbers in the naming of class rules, letters had to be used which messes up the system for selecting and saving the opacity variable.