Translation in proper Italian (lines 3 - 31)
I proposed a proper translatio of every line until line 31 (for now). Some of them are unaltered from english (that's because we use the same words in italian and we don't translate them). This is a temporary solution because I wanna be sure the translations are ok within their context.

Source: Italian guy from reddit