Change timer order list and prevent doubleclicks
-Added so timers can be sorted by name and the date to fire (and also
inverse of those)
-Added safety measure to prevent some debug options being pressed twice
such as FireAll and FixMissingTimers. Fix missing timers isn't really
major but, pressing FireAll twice can be very bad (I did it and had over
10,000 entries being written to DB)
This just adds a check to see if it's been launched already, the check
is cleared on reload or if navigating away from the timers page and
going back to it.