[Test] Different method of doing settings pages.
Normal method uses constant getSetting calls and duplicate HTML for
every button and description and glyph. This method just uses the 1 that
handles both scenarios, enable and disable so to speak.  I was also
going to do the var display = this method but that requires adding
display.everything to the HTML and ideally i'd like to get names small
instead of getSetting('topsites.enabled') ? this : this , topSites ?
this : this is better imo and the variables are already cached instead
of always calling SettingsService.