Language and TraktTV page fixes/changes
Made the checklists on the language page line up better
Updated some of the Trakt settings buttons to the other method and fixed
their appearance in the series-list view.
Changed the text on the Import from Trakt button as it was too long,
will require translations. Oh and added the spinner gif for when logging
into Trakt.tv https://www.youtube.com/watch?v=6fLxiajrL28