Some changes to the bg image on sidepanel, thoughts?
Changed from spanning the image across both panels while only 1 was open
will now display better in the first panel and expand once the second
panel opens.
Also added a 5px blur over the image. Was just a test because the the
image can look a bit distorted because of the size and opacity + bgcolor
overlapping. I think the blur kinda fixes it.