Fix up poster badges and improved SerieHeader to handle Season posters as well.
Improved the new badges on posters re: 460
Changed the poster hover effect CSS a little for better general
compaibility.
Altered how serieheader works to allow it to also work with
season.poster. This was also done to be compatible with the poster hover
effects without requiring dupe CSS.
Also removed a couple of old CSS rules.