Overhaul Posters again and allow disabling Sexy Poster Transitions
Redid the CSS and HTML for the posters yet again so that they still
display correctly if SeriesGrid is disabled.
Allowed disabling of SeriesGrid because it can drain performance
depending on your system specs
Also changed LazyBackground around a bit so we can better utilize the
img-loading and img-error classes as they weren't much use on the
element before.
Also if an image as failed to load, force show the title so we can still
see what it is