Vague commit about CSS updates
Lets see.
Added a loading spinner to the tii.me calculator
Added a .small class for loading spinner and added it in the showFiles
and tii.me spinners.
Small change to poster css making .poster a globalish normal poster size
to be referenced.
Used the new .poster css in the series-details image
Added a poster.large class that is currently used in the series-details
image
Fixed the series-details img badge position.
Fixed missing margin in theNext/Prev episode in the series-overview
where if there was no Next episode there was no margin between it and
the button below