fix standalone localStorage issue
I don't understand why this was happening. But despite a clean start up
log, subsequently, later, the app.standalone.js would start issuing
`Uncaught TypeError: Cannot read property 'getItem' of null` errors
whenever it tried accessing winState or preWinState. Perhaps some sort
of lockout occurred or I don't know what or why.
Anyway the solution (in so far as I've not had an re-occurrence again
yet) appears to be to stop reading and writing to the localStorage as
often, and cache the required variables and only update when necessary.
From a coding perspective I suppose this is best practice anyhow. So
there.