drop the reload after connection is established.
Doing a Reset immediately after the Test and Save is not good practice.
The reset breaks the connection you just tried to establish by pressing
the Test button, and destroys any reports of either success or failure
in the logs.
The user is left wondering what just happened, as no positive or
negative indicators show the user if the test/save was successful or a
failure.