Test out more responsive fastSearch re: #537
Removed the poster transforms.
Changed episodes show to 9 from 10.
Condensed the headings
Added loading indicators which are mostly useless for the series and
episodes as they load almost instantly.
When one of the sections is now empty (series, trakt, episodes) it now
transitions out of view.
Cleared the searchResults when query is empty and show message prompting
to enter query.