implement #441 for WEB torrent clients
When viewing a downloaded episode's details, DuckieTV looks for the
torrent with the connected WEB TorrentClient.
If the torrent is not found then the episode hash is removed from the
DB, which in turn prevents torrent controls from being displayed.
todo: uTorrent API. Its a Kracken and the monster has turned my brain to
mush. leaving this one for another day.