renew Trakt.TV token
added a renew-token button to the trakt.tv settings, so you can refresh
the token when it has expired without going through the username/pin
login cycle.
Tried adding the renewToken to the regular Trakt.TV traffic but the 401
error status is trapped by the $httpinterceptor so the failure code does
not get called in traktTV2. Therefore that part of the commit is
untested.