Update trakt trending cache and limit it to 350
Updated the trakt trending cache because it was pretty old and most of
the image links didnt work. I also changed the amount we fetch to 350
from 500. Mostly because we only show the first 75 and I dont think that
many people actually use the Show More option. Also with this many it
probably returns a lot of crap shows like the ones we've been recieving
and are causing the genre issues with the trending list not loading.

Note I didn't change the filename because I wasn't sure if it would
affect the build process, there is a gulp file in the build folder that
mentions the file but I thought the build system would need to be
manually updated so I just left it.