fix #752
idiot! assumed that because 1.1.3 backups did not have series custom
settings, the use of the serie.extend would be safe to use for both
1.1.3 and 1.1.4 backup imports.
Forgot that in 1.1.3 backups, if the series had episodes, then data[0]
would contain the first episodes' tvdb_id, and this resutls in it
overwriting the series own tbdb_id. duh!