also refresh the fanArt image links when user specifically asks for a serie refresh
When the user clicks on the series-details refresh-button, or the
refresh-all button on the settings-backup panel, we now also fetch new
FanArt image links rather than use the ones in the fanart DB.
This allows for the case where fanArt is now available a month or more
after  the time a brand new series was added from trakt.tv.

Should we add a periodic automated fanart image link refresh function?