bypass for problem #882
by dropping defaultValues in CRUD.enitites now that we are using the
background Adaptor.

We have coded `watched INTEGER DEFAULT 0` on the CREATE statement so the
defaultValues set is redundant here anyways.
Also in this commit, albeit unrelated to this issue, tidy up the
following:
- corrent parm order for episode.markWatched calls by
markSeriesAsWatched and markSeasonAsWatched, rootscope is the second
parm!
- use episode.isWatched() instead of hardcoded reliance on truthy
watched=0 or 1