support mixedCase font display (basically an option to turn off bebas) #972
For those of us that prefer viewing all our text in a mixed case mode.
The default is to display using Bebas.

Also helps with displaying Greek since bebas does not support that
language.

The setting on the settings,display side-panel, under the MixedCase
section.

May require further tweeking where I've missed something, but as a
proof-of-concept it seems to work \\o/