Change how sidepanel episodes work to always respect the setting.
Previously the series sidepanel episodes button settings that takes you to the latest season or current unwatched season only worked when clicking the episodes button on the series details but not on the episodes details (which you get to by clicking an episode).

I use this button a lot to update the episodes however it always goes to the latest season which is very annoying and I always thought that setting was meant to affect that button.

This does make some slight changes to translations if you wanted to update them @garfield69 though it's not a big deal.