Refactor image loading to use TMDB #1140
This switches all image loading to TMDB and lets us get show, season and episode images once again.

This does not migrate any existing data! In order to get new images you either have to wait for the show to update, manually refresh the show, or create a back up and then restore it. The images on the Trending page will also not update the trending data updates.